pyproject.toml
[project]
name = "mcp-gsheet"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-api-python-client>=2.166.0",
    "google-auth>=2.38.0",
    "google-auth-httplib2>=0.2.0",
    "mcp[cli]>=1.6.0",
]
